requirements.txt
langchain
openai
python-dotenv
pypdf
pytesseract
Pillow
python-docx
sqlalchemy
streamlit
requests